feat: Migrate PouchDB adapter
This asks the user having an old Pouch adapter (e.g. idb, cordova-sqlite)
if he wants to migrate it. We detect such need by an entry in the
localstorage set by cozy-pouch-link.

If the user refuses, we do not migrate and the user can use its Drive
normally. We ask the user each time it loads Drive until he accepts.
Once he does, we run the migration; at the end he can use again its
Drive, with better performances.

Indeed, the  indexedDB adapter greatly improves performances
for queries, replication and so on.
For example we measured the following on a Cozy with 10K docs:

* idb adapter
  * ~8s to index docs
  * ~260ms to query 1K docs
* indexeddb adapter
  * ~2.8s to index docs
  * ~160ms to query 1K docs

See also ttps://github.com/cozy/cozy-client/pull/989